Annotate sanitized InvariantErrors with numeric error codes.
Inspired by https://github.com/apollographql/apollo-client/issues/4519.

Given the { errorCodes: true } option, rollup-plugin-invariant will now
transform

  invariant(condition, message)

into

  process.env.NODE_ENV === 'production'
    ? invariant(condition, <number>)
    : invariant(condition, message)

and

  new InvariantError(message)

into

  process.env.NODE_ENV === 'production'
    ? new InvariantError(<number>)
    : new InvariantError(message)

so that production errors can be traced back to the place in the
unminified source code where they were thrown. Since the full development
version of the error is always colocated with the production one, it
should be relatively easy to determine the nature of the error.

The message for these sanitized errors will look like this:

  Invariant Violation: <number> (see https://github.com/apollographql/invariant-packages)

The provided URL is appended by the InvariantError constructor within the
ts-invariant package, so it will not be repeated at every call site.